+++ src/PhoneVerification.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Syriable\PhoneVerification;
-
-class PhoneVerification {}